backend/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
onnx>=1.14.0
opencv-python>=4.8.0
numpy>=1.24.0
tqdm>=4.65.0
pillow>=9.4.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
seaborn>=0.12.2